Reparaciones y cambios en el modulo de lkcoins #64

+++ users/views_linkcoins.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
 from rrhh.models import EntregaMoneda, datosusuario, CanjeMonedas, PremiosMonedas, MonedaLink
 from statistics import mode
-from .functions_linkcoins import estadisticasLinkcoin, email_canje_rrhh, email_canje_usuario
+from .functions_linkcoins import estadisticasLinkcoin, email_canje_rrhh, email_canje_usuario,aviso_regalo_link
 from datetime import date, datetime
 import smtplib
 from email.mime.text import MIMEText
@@ -14,70 +14,63 @@ from .functions import saludo
 def canjerealizados(request):
 
     # Listado de los datos de los premios
-
     context = {}
     context['mensaje_bievenida'] = saludo().format(request.user.first_name)
     context['data_linkcoins'] = estadisticasLinkcoin()
 
     if request.method == 'POST':
+        datos=request.POST.dict()
 
-        try:
-
-            canje = CanjeMonedas.objects.get(id = int(request.POST['ENTREGADO']))
+        if 'ENTREGADO' in datos:
+            canje = CanjeMonedas.objects.get(id = int(datos['ENTREGADO']))
             canje.entregado = "SI"
             canje.save()
 
-        except:
+            return redirect('Canjes realizados')
 
-            number = 1
-
-            vueltas = int(request.POST['cantidad'])
-
-            for i in range(vueltas):
-
-                nuevas_monedas = MonedaLink(
-                    nombre = str(date.today()) + "LINK" + str(datosusuario.objects.get(id = request.POST['usuario']).identificacion) + str(number),
-                    usuario_portador = datosusuario.objects.get(identificacion = request.user.username),
-                    fecha = date.today(),
-                    tipo = "REGALO DE LINK"
-                )
-
-                nuevas_monedas.save()
-
-                number += 1
-
-                entrega = EntregaMoneda(
-                    moneda = nuevas_monedas,
-                    fecha = date.today(),
-                    usuario_recibe = datosusuario.objects.get(id = request.POST['usuario']),
-                    mensaje = request.POST['mensaje']
-
-                )
-
-                entrega.save()
-
-
-    data_generador = []
-
-    usuarios_recibieron_generador = EntregaMoneda.objects.filter(moneda__nombre__icontains = "LINK").values_list("usuario_recibe__identificacion", flat = True).distinct()
-
-    for user in usuarios_recibieron_generador:
-        mommentos = EntregaMoneda.objects.filter(moneda__nombre__icontains = "LINK", usuario_recibe__identificacion = user).values_list("mensaje", flat = True).distinct()
-        for m in mommentos:
-            cantidad = len(EntregaMoneda.objects.filter(moneda__nombre__icontains = "LINK", usuario_recibe__identificacion = user, mensaje = m))
-            data_generador.append((datosusuario.objects.get(identificacion = user), cantidad, m))
-
-    
     
     context['data'] = CanjeMonedas.objects.all().order_by("entregado")
-    context['list_usuarios'] = datosusuario.objects.all().order_by("identificacion").exclude(estado = "NO ACTIVO")
-    context['data_generador'] = data_generador
+
 
 
     return render(request, "linkcoins/canjesrealizados.html", context)
 
 def generador_linkcoins(request):
 
+    if request.method == 'POST':
+        datos=request.POST.dict()
+
+        if 'generar' in datos:
+            cant = int(datos['cantidad']) + 1
+            mens = datos['mensaje']
+    
+            destino = datosusuario.objects.get(id = datos['usuario'])
+            for i in range(1,cant):
+
+                nueva_moneda = MonedaLink(
+                    nombre = str(date.today()) + "LINK" + str(datosusuario.objects.get(id = datos['usuario']).identificacion) + str(i),
+                    usuario_portador = datosusuario.objects.get(identificacion = request.user.username),
+                    fecha = date.today(),
+                    tipo = "REGALO DE LINK"
+                )
+                nueva_moneda.save()
+
+                entrega = EntregaMoneda(
+                    moneda = nueva_moneda,
+                    fecha = date.today(),
+                    usuario_recibe = destino,
+                    mensaje = mens,
+
+                )
+                entrega.save()
+
+            try:
+                cant=cant-1
+                aviso_regalo_link(destino.email,cant,mens)
+            except:
+                mensaje='No se pudo enviar el email'
+                
+            return redirect('Generador')
     data_generador = []
 
     usuarios_recibieron_generador = EntregaMoneda.objects.filter(moneda__nombre__icontains = "LINK").values_list("usuario_recibe__identificacion", flat = True).distinct()
@@ -93,6 +86,7 @@ def generador_linkcoins(request):
     context = {}
     context['data_generador'] = data_generador
     context['mensaje_bievenida'] = saludo().format(request.user.first_name)
+    context['list_usuarios'] = datosusuario.objects.all().order_by("identificacion").exclude(estado = "NO ACTIVO")
 
     return render(request, "linkcoins/generador_linkcoins.html", context)
 
@@ -102,21 +96,30 @@ def canjemonedas(request):
 
     context = {}
 
+    canjes=CanjeMonedas.objects.all()
+    premios=PremiosMonedas.objects.all()
+
     usuario = datosusuario.objects.get(identificacion = request.user)
 
-    monedas_recibidas = len(EntregaMoneda.objects.filter(usuario_recibe = usuario))
+    monedas_recibidas = EntregaMoneda.objects.filter(usuario_recibe = usuario).count()
 
-    monedas_canjear = monedas_recibidas - sum(CanjeMonedas.objects.filter(usuario = usuario).values_list("monedas", flat=True))
+    monedas_canjear = monedas_recibidas - sum(canjes.filter(usuario = usuario).values_list("monedas", flat=True))
 
+    mensaje = ''
+    
     if request.method == 'POST':
+        datos=request.POST.dict()
+     
+        
 
-        try:
+        today = date.today()
 
-            today = date.today()
+        if 'premio' in datos:
+            premio=datos['premio']
 
-            if today.day <= 10:
+            if today.day <= 20:
 
-                premio_solicitado = PremiosMonedas.objects.get(id = int(request.POST['premio']))
+                premio_solicitado = premios.get(id = int(premio))
 
                 if monedas_canjear >= premio_solicitado.cantidad:
 
@@ -136,69 +139,61 @@ def canjemonedas(request):
                     context['canje_realizado'] = True 
 
                 else:
-                    context['mensaje'] = "No tienes suficientes monedas para canjear el premio seleccionado"
+                    mensaje="No tienes suficientes monedas para canjear el premio seleccionado"
+              
+  
 
             else:
+                mensaje="Solo puedes canjear hasta el dia 10 de cada mes"
+  
 
-                context['mensaje'] = "Solo puedes canjear hasta el dia 10 de cada mes"
+        
 
-        except:
-
-            pass
-
-        try:
-
-            if request.POST['id'] != "0":
-                premio = PremiosMonedas.objects.get(id = int(request.POST['id']))
-                premio.nombre = request.POST['nombre']
-                premio.cantidad = request.POST['cantidad']
+        if  'id' in datos:
+            
+            if datos['id'] != "0":
+                premio = premios.get(id = int(datos['id']))
+                premio.nombre = datos['nombre']
+                premio.cantidad = datos['cantidad']
                 premio.save()
 
-            
+        
 
             else:
                 b = PremiosMonedas(
-                    nombre = request.POST['nombre'],
-                    cantidad = int(request.POST['cantidad']),
+                    nombre = datos['nombre'],
+                    cantidad = int(datos['cantidad']),
                 )
 
                 b.save()
 
-        except:
+        
 
-            pass
+        if 'borrar' in datos:
+            premio=datos['borrar']
 
-        try:
-
-            premio = PremiosMonedas.objects.get(id = int(request.POST['borrar']))
+            premio = premios.get(id = int(premio))
             premio.delete()
 
 
-        except:
 
-            pass
-
-
-    premios = PremiosMonedas.objects.all().order_by("nombre")
+    premios = premios.order_by("nombre")
 
     monedas = MonedaLink.objects.filter(usuario_portador = usuario)
 
-    monedas_recibidas = len(EntregaMoneda.objects.filter(usuario_recibe = usuario))
+    monedas_recibidas = EntregaMoneda.objects.filter(usuario_recibe = usuario).count()
 
-    monedas_canjear = monedas_recibidas - sum(CanjeMonedas.objects.filter(usuario = usuario).values_list("monedas", flat=True))
-
-    monedas_disponibles = 0
+    monedas_canjear = monedas_recibidas - sum(canjes.filter(usuario = usuario).values_list("monedas", flat=True))
     
-    for m in monedas:
+    
+    monedas_disponibles = sum([1 for m in monedas if EntregaMoneda.objects.filter(moneda = m).count() == 0])
 
-        if len(EntregaMoneda.objects.filter(moneda = m)) == 0:
-
-            monedas_disponibles += 1
 
     dato_monedas = {'monedas_recibidas':monedas_recibidas, 'monedas_disponibles':monedas_disponibles, 'monedas':monedas, 'monedas_canjear':monedas_canjear}
   
     context['premios'] = premios
     context['dato_monedas'] = dato_monedas
+    context['mensaje'] = mensaje
 
     return render(request, "linkcoins/canjemonedas.html", context)
